+++ crates/matrix-sdk/src/notification_settings/mod.rs
@@ -322,15 +322,12 @@ impl NotificationSettings {
                 // insert a `Room` rule which notifies
                 (RuleKind::Room, Notify::All)
             }
-            RoomNotificationMode::MentionsAndKeywordsOnly {
-                #[cfg(feature = "unstable-msc3768")]
-                    notify_in_app: false,
-            } => {
+            RoomNotificationMode::MentionsAndKeywordsOnly => {
                 // insert a `Room` rule which doesn't notify
                 (RuleKind::Room, Notify::None)
             }
             #[cfg(feature = "unstable-msc3768")]
-            RoomNotificationMode::MentionsAndKeywordsOnly { notify_in_app: true } => {
+            RoomNotificationMode::MentionsAndKeywordsOnlyTheRestInApp => {
                 // insert a `Room` rule which notifies in-app only
                 (RuleKind::Room, Notify::InAppOnly)
             }
@@ -722,10 +719,7 @@ mod tests {
         let settings = from_insert_rules(&client, vec![(RuleKind::Room, &room_id, false)]);
         assert_eq!(
             settings.get_user_defined_room_notification_mode(&room_id).await.unwrap(),
-            RoomNotificationMode::MentionsAndKeywordsOnly {
-                #[cfg(feature = "unstable-msc3768")]
-                notify_in_app: false
-            }
+            RoomNotificationMode::MentionsAndKeywordsOnly
         );
     }
 
@@ -779,10 +773,7 @@ mod tests {
         let settings = NotificationSettings::new(client.to_owned(), ruleset.to_owned());
         assert_eq!(
             settings.get_default_room_notification_mode(IsEncrypted::No, IsOneToOne::Yes).await,
-            RoomNotificationMode::MentionsAndKeywordsOnly {
-                #[cfg(feature = "unstable-msc3768")]
-                notify_in_app: false
-            }
+            RoomNotificationMode::MentionsAndKeywordsOnly
         );
 
         // The default mode must be `MentionsAndKeywords` if the corresponding Underride
@@ -794,10 +785,7 @@ mod tests {
         let settings = NotificationSettings::new(client, ruleset);
         assert_eq!(
             settings.get_default_room_notification_mode(IsEncrypted::No, IsOneToOne::Yes).await,
-            RoomNotificationMode::MentionsAndKeywordsOnly {
-                #[cfg(feature = "unstable-msc3768")]
-                notify_in_app: false
-            }
+            RoomNotificationMode::MentionsAndKeywordsOnly
         );
     }
 
@@ -939,12 +927,9 @@ mod tests {
 
         let new_modes = [
             RoomNotificationMode::AllMessages,
-            RoomNotificationMode::MentionsAndKeywordsOnly {
-                #[cfg(feature = "unstable-msc3768")]
-                notify_in_app: false,
-            },
+            RoomNotificationMode::MentionsAndKeywordsOnly,
             #[cfg(feature = "unstable-msc3768")]
-            RoomNotificationMode::MentionsAndKeywordsOnly { notify_in_app: true },
+            RoomNotificationMode::MentionsAndKeywordsOnlyTheRestInApp,
             RoomNotificationMode::Mute,
         ];
         for new_mode in new_modes {
@@ -1121,10 +1106,7 @@ mod tests {
         let settings = from_insert_rules(&client, vec![(RuleKind::Room, &room_id, false)]);
         assert_eq!(
             settings.get_user_defined_room_notification_mode(&room_id).await.unwrap(),
-            RoomNotificationMode::MentionsAndKeywordsOnly {
-                #[cfg(feature = "unstable-msc3768")]
-                notify_in_app: false
-            }
+            RoomNotificationMode::MentionsAndKeywordsOnly
         );
 
         // Unmute the room
@@ -1133,10 +1115,7 @@ mod tests {
         // The ruleset must not be modified
         assert_eq!(
             settings.get_user_defined_room_notification_mode(&room_id).await.unwrap(),
-            RoomNotificationMode::MentionsAndKeywordsOnly {
-                #[cfg(feature = "unstable-msc3768")]
-                notify_in_app: false
-            }
+            RoomNotificationMode::MentionsAndKeywordsOnly
         );
 
         let room_rules = get_custom_rules_for_room(&settings, &room_id).await;
@@ -1233,10 +1212,7 @@ mod tests {
             .set_default_room_notification_mode(
                 IsEncrypted::No,
                 IsOneToOne::No,
-                RoomNotificationMode::MentionsAndKeywordsOnly {
-                    #[cfg(feature = "unstable-msc3768")]
-                    notify_in_app: false,
-                },
+                RoomNotificationMode::MentionsAndKeywordsOnly,
             )
             .await
             .unwrap();
@@ -1258,10 +1234,7 @@ mod tests {
         // reflect the change.
         assert_matches!(
             settings.get_default_room_notification_mode(IsEncrypted::No, IsOneToOne::No).await,
-            RoomNotificationMode::MentionsAndKeywordsOnly {
-                #[cfg(feature = "unstable-msc3768")]
-                notify_in_app: false
-            }
+            RoomNotificationMode::MentionsAndKeywordsOnly
         );
     }
 
@@ -1300,10 +1273,7 @@ mod tests {
             .set_default_room_notification_mode(
                 IsEncrypted::No,
                 IsOneToOne::Yes,
-                RoomNotificationMode::MentionsAndKeywordsOnly {
-                    #[cfg(feature = "unstable-msc3768")]
-                    notify_in_app: false,
-                },
+                RoomNotificationMode::MentionsAndKeywordsOnly,
             )
             .await
             .unwrap();
@@ -1325,10 +1295,7 @@ mod tests {
         // reflect the change.
         assert_matches!(
             settings.get_default_room_notification_mode(IsEncrypted::No, IsOneToOne::Yes).await,
-            RoomNotificationMode::MentionsAndKeywordsOnly {
-                #[cfg(feature = "unstable-msc3768")]
-                notify_in_app: false
-            }
+            RoomNotificationMode::MentionsAndKeywordsOnly
         );
     }
 
@@ -1650,10 +1617,7 @@ mod tests {
             .set_default_room_notification_mode(
                 IsEncrypted::No,
                 IsOneToOne::No,
-                RoomNotificationMode::MentionsAndKeywordsOnly {
-                    #[cfg(feature = "unstable-msc3768")]
-                    notify_in_app: false,
-                },
+                RoomNotificationMode::MentionsAndKeywordsOnly,
             )
             .await
             .unwrap();
@@ -1662,10 +1626,7 @@ mod tests {
         // reflect the change.
         assert_matches!(
             settings.get_default_room_notification_mode(IsEncrypted::No, IsOneToOne::No).await,
-            RoomNotificationMode::MentionsAndKeywordsOnly {
-                #[cfg(feature = "unstable-msc3768")]
-                notify_in_app: false
-            }
+            RoomNotificationMode::MentionsAndKeywordsOnly
         );
     }